refactor: Update useFetchPost hook to usePostFetch
